backend/requirements.txt
aiosmtplib==3.0.2
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
blinker==1.9.0
certifi==2025.6.15
click==8.2.1
colorama==0.4.6
dependencies==7.7.1
Deprecated==1.2.18
dnspython==2.7.0
email_validator==2.2.0
factory_boy==3.3.3
Faker==37.4.0
fastapi==0.115.13
fastapi-cli==0.0.7
fastapi-mail==1.5.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.4.0
Mako==1.3.10
mando==0.7.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.10.18
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pytest==8.4.1
pytest-asyncio==1.0.0
python-dotenv==1.1.0
python-multipart==0.0.20
python-telegram-bot==22.1
PyYAML==6.0.2
radon==6.0.1
rich==14.0.0
rich-toolkit==0.14.7
ruff==0.12.0
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.10.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.2
